@@ -1,14 +1,14 @@
 plugins {
 	id 'java'
-	id 'org.springframework.boot' version '2.7.17-SNAPSHOT'
-	id 'io.spring.dependency-management' version '1.0.15.RELEASE'
+	id 'org.springframework.boot' version '3.1.5'
+	id 'io.spring.dependency-management' version '1.1.3'
 }
 
 group = 'com.client'
 version = '0.0.1-SNAPSHOT'
 
 java {
-	sourceCompatibility = '11'
+	sourceCompatibility = '17'
 }
 
 configurations {
@@ -19,26 +19,22 @@ configurations {
 
 repositories {
 	mavenCentral()
-	maven { url 'https://repo.spring.io/milestone' }
-	maven { url 'https://repo.spring.io/snapshot' }
 }
 
 ext {
-	set('springBootAdminVersion', "2.3.0")//"3.1.5") //"2.3.0") //añadido
-	set('springCloudVersion', "2021.0.8")
+	set('springBootAdminVersion', "3.1.5")
+	set('springCloudVersion', "2022.0.4")
 }
 
 dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-web'		//
-	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'	//
-	implementation 'org.springframework.boot:spring-boot-starter-validation'//
-	implementation 'org.springframework.boot:spring-boot-starter-actuator'	//
-
-	implementation 'de.codecentric:spring-boot-admin-starter-client'		//añadido
-	implementation 'org.springframework.cloud:spring-cloud-starter-config'	//
-
-	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client' //
-	implementation 'org.springframework.cloud:spring-cloud-starter-sleuth'	//añadido
+	implementation 'org.springframework.boot:spring-boot-starter-actuator'
+	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+	implementation 'org.springframework.boot:spring-boot-starter-validation'
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	implementation 'de.codecentric:spring-boot-admin-starter-client'
+	implementation 'org.springframework.cloud:spring-cloud-starter-config'
+	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
+	//implementation 'org.springframework.cloud:spring-cloud-starter-sleuth' //añadido
 
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
@@ -49,7 +45,7 @@ dependencies {
 
 dependencyManagement {
 	imports {
-		mavenBom "de.codecentric:spring-boot-admin-dependencies:${springBootAdminVersion}" //añadido
+		mavenBom "de.codecentric:spring-boot-admin-dependencies:${springBootAdminVersion}"
 		mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
 	}
 }
